Add information about minimum screen resolution
After some deliberation within the development team we felt it would be safe to say that: Micetro Web UI supports a minimum browser viewport area of 1366x768 pixels when using a maximum 100% zoom setting. Lower resolutions are not supported.

This is a screen resolution we are fairly comfortable with, we should be able to gracefully handle this resolution. Any lower resolution and we might start to get some issue related to the overall layout which could force us into extra work later on.

As part of this pull-request some minor layout changes were done to make the browser section more clear